10: Added example scene, some small fixes

+++ include/scene.h
@@ -28,7 +28,7 @@ class PointerManager {
 
 struct Scene {
     Object** objects;
-    int number_of_objects;
+    size_t number_of_objects;
     Camera camera;
     PointerManager* pointer_manager;
     Medium* medium;
